+++ CppScripts/BBBhw/GPIO.h
@@ -119,7 +119,7 @@ private:// Variables
 	int duration_FinalInitialMeasTrigAuxArray[MaxNumPulses]={0};
 	int duration_FinalInitialMeasTrigAuxAvg=0;
 	unsigned long long int TrigAuxIterCount=0;
-	unsigned long SynchTrigPeriod=65536; //For histogram analysis this is 65536
+	double SynchTrigPeriod=65536; //For histogram analysis this is 65536
 	unsigned long TimeClockMarging=100000;// In nanoseconds
 	unsigned long long int TimeClockMargingExtra=10*TimeClockMarging;// In nanoseconds
 	unsigned long long int TimeElpasedNow_time_as_count=0;
